pyproject.toml
[project]
name = "modulo-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "django-environ>=0.11.0",
    "psycopg2-binary>=2.9.11",
    "django-filter>=25.2",
    "django-jazzmin>=3.0.1",
    "drf-api-logger>=1.1.21",
    "django-unfold>=0.69.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
]
prod = [
    "gunicorn>=23.0.0",
]
